+++ .gitignore
@@ -26,8 +26,8 @@ __pycache__/
 docs/source/developer/autogen*
 
 # Ignore generated cpp files
-numba_dppy/dpnp_glue/*.cpp
-numba_dppy/dpnp_glue/*.h
+numba_dppy/dpnp_iface/*.cpp
+numba_dppy/dpnp_iface/*.h
 
 # Ignore coverage results
 .coverage
