+++ numba_dppy/dpnp_iface/dpnp_array_creations_impl.py
@@ -19,8 +19,8 @@ from numba.core.typing import signature
 
 import numba_dppy
 import numba_dppy.dpctl_iface as dpctl_functions
-import numba_dppy.dpnp_glue as dpnp_lowering
-import numba_dppy.dpnp_glue.dpnpimpl as dpnp_ext
+import numba_dppy.dpnp_iface as dpnp_lowering
+import numba_dppy.dpnp_iface.dpnpimpl as dpnp_ext
 
 from . import stubs
