+++ numba_dppy/dpnp_iface/dpnp_logic.py
@@ -18,8 +18,8 @@ from numba.core.extending import overload
 from numba.core.typing import signature
 
 import numba_dppy.dpctl_iface as dpctl_functions
-import numba_dppy.dpnp_glue as dpnp_lowering
-import numba_dppy.dpnp_glue.dpnpimpl as dpnp_ext
+import numba_dppy.dpnp_iface as dpnp_lowering
+import numba_dppy.dpnp_iface.dpnpimpl as dpnp_ext
 
 from . import stubs
